foodgramm/requirements.txt
asgiref==3.8.1
Django==4.2.11
djangorestframework==3.15.1
drf-yasg==1.21.7
flake8==7.0.0
inflection==0.5.1
mccabe==0.7.0
packaging==24.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.12.1
uritemplate==4.1.1
